Use one mnemonic phrase for keystore (#950)
* draft implements one mnemonic phrase

* complete mnemonic phrase encrypt and reuse

* fixup ci check

* fixup error

* fixup some testcase

* cleanup account import and update cli

* fixup integration test case

* remove encrypt-keystore option in cli
