+++ crates/rooch-framework-tests/src/tests/native_validator_tests.rs
@@ -2,7 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use moveos_types::transaction::MoveAction;
-use rooch_key::keystore::{AccountKeystore, InMemKeystore};
+use rooch_key::keystore::account_keystore::AccountKeystore;
+use rooch_key::keystore::memory_keystore::InMemKeystore;
 use rooch_types::framework::empty::Empty;
 use rooch_types::transaction::{rooch::RoochTransactionData, AbstractTransaction};
