+++ crates/rooch-framework-tests/src/tests/transaction_validator_tests.rs
@@ -9,7 +9,8 @@ use move_core_types::value::MoveValue;
 use move_core_types::vm_status::{AbortLocation, VMStatus};
 use moveos_types::move_types::FunctionId;
 use moveos_types::{module_binding::ModuleBinding, transaction::MoveAction};
-use rooch_key::keystore::{AccountKeystore, InMemKeystore};
+use rooch_key::keystore::account_keystore::AccountKeystore;
+use rooch_key::keystore::memory_keystore::InMemKeystore;
 use rooch_types::address::MultiChainAddress;
 use rooch_types::framework::session_key::SessionKeyModule;
 use rooch_types::framework::timestamp::TimestampModule;
