+++ crates/rooch-key/src/key_derive.rs
@@ -18,6 +18,7 @@ use rooch_types::error::RoochError;
 use rooch_types::key_struct::{EncryptionData, GenerateNewKeyPair, GeneratedKeyPair};
 use rooch_types::multichain_id::RoochMultiChainID;
 use slip10_ed25519::derive_ed25519_private_key;
+use std::str::FromStr;
 use std::string::String;
 
 // Purpose constants
@@ -26,10 +27,7 @@ pub const DERIVATION_PATH_PURPOSE_SCHNORR: u32 = 44;
 pub const DERIVATION_PATH_PURPOSE_ECDSA: u32 = 54;
 pub const DERIVATION_PATH_PURPOSE_SECP256R1: u32 = 74;
 
-pub fn encrypt_private_key(
-    private_key: &[u8],
-    password: Option<String>,
-) -> Result<EncryptionData, RoochError> {
+pub fn encrypt_key(key: &[u8], password: Option<String>) -> Result<EncryptionData, RoochError> {
     let nonce = ChaCha20Poly1305::generate_nonce(&mut OsRng);
     let mut output_key_material = [0u8; 32];
     Argon2::default()
@@ -43,7 +41,7 @@ pub fn encrypt_private_key(
     let cipher = ChaCha20Poly1305::new_from_slice(&output_key_material)
         .map_err(|e| RoochError::KeyConversionError(e.to_string()))?;
 
-    let ciphertext_with_tag = match cipher.encrypt(&nonce, private_key) {
+    let ciphertext_with_tag = match cipher.encrypt(&nonce, key) {
         Ok(ciphertext) => ciphertext,
         Err(_) => {
             return Err(RoochError::KeyConversionError(
@@ -62,7 +60,7 @@ pub fn encrypt_private_key(
     })
 }
 
-pub fn decrypt_private_key(
+pub fn decrypt_key(
     nonce: &[u8],
     ciphertext: &[u8],
     tag: &[u8],
@@ -125,7 +123,7 @@ pub fn derive_private_key_from_path(
     seed: &[u8],
     derivation_path: Option<DerivationPath>,
 ) -> Result<Vec<u8>, RoochError> {
-    let path = validate_path(derivation_path)?;
+    let path = validate_derivation_path(derivation_path)?;
     let indexes = path.iter().map(|i| i.into()).collect::<Vec<_>>();
     let derived = derive_ed25519_private_key(seed, &indexes);
     let sk = Ed25519PrivateKey::from_bytes(&derived)
@@ -153,7 +151,7 @@ pub fn retrieve_key_pair(
     let tag = Base64::decode(&encryption.tag)
         .map_err(|e| RoochError::KeyConversionError(e.to_string()))?;
 
-    let private_key = decrypt_private_key(&nonce, &ciphertext, &tag, password)?;
+    let private_key = decrypt_key(&nonce, &ciphertext, &tag, password)?;
 
     let kp = Ed25519KeyPair::from(
         Ed25519PrivateKey::from_bytes(&private_key)
@@ -163,55 +161,89 @@ pub fn retrieve_key_pair(
     Ok(kp.into())
 }
 
-pub fn validate_path(path: Option<DerivationPath>) -> Result<DerivationPath, RoochError> {
+pub fn validate_derivation_path(
+    path: Option<DerivationPath>,
+) -> Result<DerivationPath, anyhow::Error> {
     let (purpose, coin_type) = (
         DERIVATION_PATH_PURPOSE_ED25519,
-        RoochMultiChainID::Sui as u32,
+        RoochMultiChainID::Rooch as u32,
     );
 
     match path {
         Some(p) => {
             if let &[p_purpose, p_coin_type, account, change, address] = p.as_ref() {
-                if p_purpose == bip32::ChildNumber(purpose)
-                    && p_coin_type == bip32::ChildNumber(coin_type)
+                if p_purpose == bip32::ChildNumber::new(purpose, true)?
+                    && p_coin_type == bip32::ChildNumber::new(coin_type, true)?
                     && account.is_hardened()
                     && change.is_hardened()
                     && address.is_hardened()
                 {
                     Ok(p)
                 } else {
-                    Err(RoochError::SignatureKeyGenError("Invalid path".to_owned()))
+                    Err(anyhow::anyhow!("Invalid derivation path: {}", p))
                 }
             } else {
-                Err(RoochError::SignatureKeyGenError("Invalid path".to_owned()))
+                Err(anyhow::anyhow!("Invalid derivation path: {}", p))
             }
         }
-        None => Ok(format!("m/{}'/{}/0'/0'/0'", purpose, coin_type)
+        None => Ok(format!("m/{}'/{}'/0'/0'/0'", purpose, coin_type)
             .parse()
-            .map_err(|_| RoochError::SignatureKeyGenError("Cannot parse path".to_owned()))?),
+            .map_err(|e| anyhow::anyhow!("Cannot parse derivation path, error:{}", e))?),
     }
 }
 
+/// Derivation path template
+/// Which tells a wallet how to derive a specific key within a tree of keys
+/// https://github.com/bitcoin/bips/blob/master/bip-0044.mediawiki
+/// for ed25529
+/// m / purpose' / coin_type' / account' / change' / address_index'
+pub fn generate_derivation_path(account_index: u32) -> Result<DerivationPath, RoochError> {
+    let (purpose, coin_type) = (
+        DERIVATION_PATH_PURPOSE_ED25519,
+        RoochMultiChainID::Rooch as u32,
+    );
+
+    DerivationPath::from_str(
+        format!("m/{}'/{}'/0'/0'/{}'", purpose, coin_type, account_index).as_str(),
+    )
+    .map_err(|_| RoochError::SignatureKeyGenError("Cannot parse derivation path".to_owned()))
+}
+
 pub fn generate_new_key_pair(
+    mnemonic_phrase: Option<String>,
     derivation_path: Option<DerivationPath>,
     word_length: Option<String>,
     password: Option<String>,
 ) -> Result<GeneratedKeyPair, anyhow::Error> {
-    let mnemonic = Mnemonic::new(parse_word_length(word_length)?, Language::English);
+    // Reuse the mnemonic phrase to derive new address
+    let mnemonic = match mnemonic_phrase {
+        Some(phrase) => {
+            //TODO need validate mnemonic phrase ?
+            Mnemonic::from_phrase(phrase.as_str(), Language::English)?
+        }
+        None => Mnemonic::new(parse_word_length(word_length)?, Language::English),
+    };
     let seed = Seed::new(&mnemonic, "");
 
     let sk = derive_private_key_from_path(seed.as_bytes(), derivation_path)?;
 
-    let encryption = encrypt_private_key(&sk, password).expect("Encryption failed for private key");
+    let private_key_encryption =
+        encrypt_key(&sk, password.clone()).expect("Encryption failed for private key");
+    let mnemonic_phrase_encryption = encrypt_key(mnemonic.phrase().as_bytes(), password)
+        .expect("Encryption failed for mnemonic phrase");
 
     let address = derive_address_from_private_key(sk)?;
 
     let result = GenerateNewKeyPair {
-        encryption,
-        mnemonic: mnemonic.phrase().to_string(),
+        private_key_encryption,
+        mnemonic_phrase_encryption,
+        mnemonic_phrase: mnemonic.phrase().to_string(),
     };
 
-    Ok(GeneratedKeyPair { address, result })
+    Ok(GeneratedKeyPair {
+        address,
+        key_pair_data: result,
+    })
 }
 
 fn parse_word_length(s: Option<String>) -> Result<MnemonicType, anyhow::Error> {
