+++ crates/rooch-types/src/key_struct.rs
@@ -1,10 +1,9 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::address::RoochAddress;
 use fastcrypto::encoding::{Base64, Encoding};
 use serde::{Deserialize, Serialize};
-
-use crate::address::RoochAddress;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct EncryptionData {
@@ -13,12 +12,13 @@ pub struct EncryptionData {
     pub tag: String,
 }
 pub struct GenerateNewKeyPair {
-    pub encryption: EncryptionData,
-    pub mnemonic: String,
+    pub mnemonic_phrase: String,
+    pub private_key_encryption: EncryptionData,
+    pub mnemonic_phrase_encryption: EncryptionData,
 }
 pub struct GeneratedKeyPair {
     pub address: RoochAddress,
-    pub result: GenerateNewKeyPair,
+    pub key_pair_data: GenerateNewKeyPair,
 }
 
 impl EncryptionData {
@@ -41,4 +41,18 @@ impl EncryptionData {
             tag: Base64::encode(tag),
         }
     }
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct MnemonicResult {
+    pub mnemonic_phrase: String,
+    pub mnemonic_phrase_key: String,
+    pub mnemonic_data: MnemonicData,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct MnemonicData {
+    // pub mnemonic_phrase: String,
+    pub addresses: Vec<RoochAddress>,
+    pub mnemonic_phrase_encryption: EncryptionData,
 }
