+++ crates/rooch/src/commands/account/commands/mod.rs
@@ -3,8 +3,6 @@
 
 pub mod balance;
 pub mod create;
-pub mod import;
 pub mod list;
 pub mod nullify;
 pub mod switch;
-pub mod update;
