+++ crates/rooch/src/commands/move_cli/commands/run_function.rs
@@ -5,7 +5,8 @@ use crate::cli_types::{ArgWithType, CommandAction, TransactionOptions, WalletCon
 use async_trait::async_trait;
 use clap::Parser;
 use moveos_types::{move_types::FunctionId, transaction::MoveAction};
-use rooch_key::{key_derive::verify_password, keystore::AccountKeystore};
+use rooch_key::key_derive::verify_password;
+use rooch_key::keystore::account_keystore::AccountKeystore;
 use rooch_rpc_api::jsonrpc_types::{ExecuteTransactionResponseView, TypeTagView};
 use rooch_types::{
     address::RoochAddress,
@@ -96,10 +97,8 @@ impl CommandAction<ExecuteTransactionResponseView> for RunFunction {
                         .sign_transaction_via_session_key(&sender, tx_data, &session_key, None)
                         .map_err(|e| RoochError::SignMessageError(e.to_string()))?
                 } else {
-                    let password = prompt_password(
-                        "Enter the password saved in client config to run functions:",
-                    )
-                    .unwrap_or_default();
+                    let password =
+                        prompt_password("Enter the password to run functions:").unwrap_or_default();
                     let is_verified = verify_password(
                         Some(password.clone()),
                         context.keystore.get_password_hash(),
@@ -127,10 +126,8 @@ impl CommandAction<ExecuteTransactionResponseView> for RunFunction {
                 if context.keystore.get_if_password_is_empty() {
                     context.sign_and_execute(sender, action, None).await
                 } else {
-                    let password = prompt_password(
-                        "Enter the password saved in client config to run functions:",
-                    )
-                    .unwrap_or_default();
+                    let password =
+                        prompt_password("Enter the password to run functions:").unwrap_or_default();
                     let is_verified = verify_password(
                         Some(password.clone()),
                         context.keystore.get_password_hash(),
